apps/text-embeddings/requirements.txt
torch>=2.0
open_clip_torch>=2.23
numpy
Pillow
